+++ tests/testthat/test-app.R
@@ -44,3 +44,45 @@ test_that("Basic functionality", {
   stopAllServers()
   expect_equal(length(listServers()), 0)
 })
+
+
+test_that("Empty and NULL headers are OK", {
+  s <- startServer("127.0.0.1", randomPort(),
+    list(
+      call = function(req) {
+        if (req$PATH_INFO == "/null") {
+          list(
+            status = 200L,
+            headers = NULL,
+            body = ""
+          )
+        } else if (req$PATH_INFO == "/emptylist") {
+          list(
+            status = 200L,
+            headers = list(),
+            body = ""
+          )
+        } else if (req$PATH_INFO == "/noheaders") {
+          list(
+            status = 200L,
+            body = ""
+          )
+        }
+      }
+    )
+  )
+  on.exit(s$stop())
+  expect_equal(length(listServers()), 1)
+
+  r <- fetch(local_url("/null", s$getPort()))
+  expect_equal(r$status_code, 200)
+  expect_identical(r$content, raw())
+
+  r <- fetch(local_url("/emptylist", s$getPort()))
+  expect_equal(r$status_code, 200)
+  expect_identical(r$content, raw())
+
+  r <- fetch(local_url("/noheaders", s$getPort()))
+  expect_equal(r$status_code, 200)
+  expect_identical(r$content, raw())
+})